Further refactoring of LibUtils. Minor cleanups.
The LibUtils class now allows placing the dependent
libraries into subdirectories in the JAR that are
named based on the OS and architecture, to avoid
name collisions. It also writes them to temporary
files that are contained in subdirectories of the
temp directory that clearly mark them as dependents
of the main library.

Minor cleanups of the POM, CMake files and readme.