Retain the event that will be used for waiting.
When a non-blocking operation with native source memory
is enqueued, the event will be used to determine when
the source object can be released for garbage collection.
But when the associated event was released manually,
before the wait operation could start, this caused a
painful crash in clWaitForEvents. Increasing the reference
count of the event with clRetainEvent will prevent this.