Ensure finding the OpenCL library "libOpenCL.so.1" on linux
It is currently required to install dev bindings for OpenCL on linux to
run JOCL. Only the dev bindings contain "libOpenCL.so". Ideally the dev
bindings would only be required during compilation but not during runtime.

The OpenCL library binaries on standard linux installation are called:
- "libOpenCL.so.1.0.0"
- "libOpenCL.so.1"        (this is a link to the first file)
The have these additional version suffixes to avoid version conflicts.

This PR changes JOCL to load "libOpenCL.so.1" on linux. This means that
JOCL can run on linux without requiring dev bindings.
(Tested on Ubuntu 24.04)